update recordatory comp
